refactor: moved Main to page folder
