+++ src/App.tsx
@@ -1,27 +1,22 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
-import Header from "./components/layout/Header"; 
-import Footer from "./components/layout/Footer"; 
-import Home from "../src/components/layout/Home";
+import Header from "./components/layout/Header";
+import Footer from "./components/layout/Footer";
+import Home from "./pages/Home";
 
 const App = () => {
   return (
     <Router>
-        <Header />
-        <main className="flex-grow">
-          <Routes>
-            <Route path="/" element={<Home />} />
-            {/* Add other routes here */}
-          </Routes>
-        </main>
-        <Footer />
+      <Header />
+      <main className="flex-grow">
+        <Routes>
+          <Route path="/" element={<Home />} />
+          {/* Add other routes here */}
+        </Routes>
+      </main>
+      <Footer />
     </Router>
   );
 };
 
 export default App;
-
-
-
-
-
